+++ jupyterlab-megane/src/filetypes.ts
@@ -119,6 +119,14 @@ export const STRUCTURE_FILETYPES_TEXT: DocumentRegistry.IFileType[] = [
     contentType: "file",
   },
   {
+    name: "megane-c3xml",
+    displayName: "Chem3D XML",
+    extensions: [".c3xml"],
+    mimeTypes: ["chemical/x-c3xml"],
+    fileFormat: "text",
+    contentType: "file",
+  },
+  {
     name: "megane-lammps-dump",
     displayName: "LAMMPS dump",
     extensions: [".lammpstrj", ".dump", ".trj"],